Added engine/tqt.*, a helper module for handling texture quadtrees.
Chunkdemo can use tqt's now!  There are a number of sampling/alignment problems which create pops and less-than-ideal textures, but those will get solved.  Pretty cool results.

TODO: Fix maketqt's performance problem and sampling problems.  Threading and unloading for chunkdemo.  Chunkdemo able to handle .tqt's of different depth than the .chu.

+++ tu-testbed/engine/cull.cpp
@@ -30,9 +30,9 @@ result_info	compute_box_visibility(const vec3& center, const vec3& extent, const
 		const plane_info&	p = frustum[i];
 		// Check box against this plane.
 		float	d = p.normal * center - p.d;
-		float	extent_toward_plane = fabs(extent.get_x() * p.normal.get_x())
-			+ fabs(extent.get_y() * p.normal.get_y())
-			+ fabs(extent.get_z() * p.normal.get_z());
+		float	extent_toward_plane = fabsf(extent.get_x() * p.normal.get_x())
+			+ fabsf(extent.get_y() * p.normal.get_y())
+			+ fabsf(extent.get_z() * p.normal.get_z());
 		if (d < 0) {
 			if (-d > extent_toward_plane) {
 				// Box is culled by plane; it's not visible.
